font size should be CGFloat
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@3948 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Addresses/Frameworks/AddressView/ADPersonView.h
@@ -39,7 +39,7 @@ extern NSString * const ADPeoplePboardType;
   BOOL _acceptsDrop;
   BOOL _noteTextChanged;
 
-  float _fontSize;
+  CGFloat _fontSize;
 
   BOOL _mouseDownOnSelf;
 }
@@ -79,8 +79,8 @@ extern NSString * const ADPeoplePboardType;
 - (void) setAcceptsDrop: (BOOL) yesno;
 - (BOOL) acceptsDrop;
 
-- (void) setFontSize: (float) fontSize;
-- (float) fontSize;
+- (void) setFontSize: (CGFloat) fontSize;
+- (CGFloat) fontSize;
 @end
 
 @interface ADPersonView (PropertyMangling)
